Sensor Time stamp data type change from 32 bit to 64 bit - Timestamp datatype changed from uint32 to unit64 in all associated functions ,prototypes and debug prints - Rounded  RTC_TICK_US_FLT macro to 3 decimal places

+++ embedded/common/app/appmsgstruct.h
@@ -45,7 +45,7 @@ typedef struct MsgSensorDataTag
     int32_t     TiltError;
 } MsgSensorData;
 
-typedef struct MsgSensorBoolTag 
+typedef struct MsgSensorBoolTag
 {
     uint64_t timeStamp;
     uint8_t  active;
@@ -65,7 +65,7 @@ typedef MsgSensorBoolData MsgStepDetData;
 
 typedef struct MsgSensorDataRdyTag
 {
-    uint32_t     timeStamp;
+    uint64_t     timeStamp;
     uint8_t      sensorId;
 } MsgSensorDataRdy;
